requirements.txt
flask==2.0.1
numpy==1.21.0
opencv-python==4.5.3.56
pillow==8.3.1
